elaborate.cxx: fix 32-bit build
Without the change the build fails on i686-linux as:

    elaborate.cxx:5119:33: error:
      format '%ld' expects argument of type 'long int',
      but argument 2 has type 'int64_t' {aka 'long long int'} [-Werror=format=]
     5119 |       session.print_warning (_F("Collapsing unresolved @define to %ld [stapprobes]", value), e->tok);
          |                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ elaborate.cxx
@@ -5116,7 +5116,7 @@ const_folder::visit_defined_op (defined_op* e)
       // Don't be greedy... we'll only collapse one at a time so type
       // resolution can have another go at it.
       relaxed_p = false;
-      session.print_warning (_F("Collapsing unresolved @define to %ld [stapprobes]", value), e->tok);
+      session.print_warning (_F("Collapsing unresolved @define to %lld [stapprobes]", (long long)value), e->tok);
       literal_number* n = new literal_number (value);
       n->tok = e->tok;
       n->visit (this);
